Big decimal

+++ src/main/java/uady/mx/nube/dto/PagoDTO.java
@@ -1,9 +1,11 @@
 package uady.mx.nube.dto;
 
+import java.math.BigDecimal;
+
 public class PagoDTO {
 
     private Integer idPago;
-    private Double monto;
+    private BigDecimal monto;
     private String cuentaOrigen;
     private String cuentaDestino;
     private String fechaRegistro;
@@ -11,7 +13,7 @@ public class PagoDTO {
     private String estado;
     private Integer retry;
 
-    public PagoDTO(Integer idPago, Double monto, String cuentaOrigen, String cuentaDestino, String fechaRegistro,
+    public PagoDTO(Integer idPago, BigDecimal monto, String cuentaOrigen, String cuentaDestino, String fechaRegistro,
             String fechaProcesa, String estado) {
         this.idPago = idPago;
         this.monto = monto;
@@ -30,7 +32,7 @@ public class PagoDTO {
         return this.idPago;
     }
 
-    public Double getMonto() {
+    public BigDecimal getMonto() {
         return this.monto;
     }
 
@@ -54,7 +56,7 @@ public class PagoDTO {
         return this.estado;
     }
 
-    public Integer getRetry(){
+    public Integer getRetry() {
         return this.retry;
     }
 
@@ -62,7 +64,7 @@ public class PagoDTO {
         this.idPago = idPago;
     }
 
-    public void setMonto(Double monto) {
+    public void setMonto(BigDecimal monto) {
         this.monto = monto;
     }
 
@@ -86,7 +88,7 @@ public class PagoDTO {
         this.estado = estado;
     }
 
-    public void setRetry(Integer retry){
+    public void setRetry(Integer retry) {
         this.retry = retry;
     }
 }
