+++ src/main/java/uady/mx/nube/model/Pago.java
@@ -1,10 +1,12 @@
 package uady.mx.nube.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.*;
 
 import uady.mx.nube.enums.EstadoEnum;
+
 @Entity
 @Table(name = "pagos")
 public class Pago {
@@ -14,14 +16,14 @@ public class Pago {
     private Integer id;
 
     @Column(name = "monto")
-    private Double monto;
+    private BigDecimal monto;
 
     @ManyToOne
-    @JoinColumn(name="cuenta_origen", referencedColumnName = "id")
+    @JoinColumn(name = "cuenta_origen", referencedColumnName = "id")
     private Cuenta cuentaOrigen;
 
     @ManyToOne
-    @JoinColumn(name="cuenta_destino", referencedColumnName = "id")
+    @JoinColumn(name = "cuenta_destino", referencedColumnName = "id")
     private Cuenta cuentaDestino;
 
     @Column(name = "fecha_registro")
@@ -34,79 +36,76 @@ public class Pago {
     @Enumerated(EnumType.STRING)
     private EstadoEnum estado;
 
-    public Pago(double monto, Cuenta cuentaOrigen, Cuenta cuentaDestino){
+    public Pago(BigDecimal monto, Cuenta cuentaOrigen, Cuenta cuentaDestino) {
         this.monto = monto;
         this.cuentaOrigen = cuentaOrigen;
         this.cuentaDestino = cuentaDestino;
     }
 
-    public Pago(){}
+    public Pago() {
+    }
 
-    public Integer getId(){
+    public Integer getId() {
         return this.id;
     }
 
-    public Double getMonto(){
+    public BigDecimal getMonto() {
         return this.monto;
     }
 
-    public Cuenta getCuentaOrigen(){
+    public Cuenta getCuentaOrigen() {
         return this.cuentaOrigen;
     }
 
-    public Cuenta getCuentaDestino(){
+    public Cuenta getCuentaDestino() {
         return this.cuentaDestino;
     }
 
-    public Date getFechaRegistro(){
+    public Date getFechaRegistro() {
         return this.fechaRegistro;
     }
 
-    public Date getFechaProcesa(){
+    public Date getFechaProcesa() {
         return this.fechaProcesa;
     }
 
-    public EstadoEnum getEstado(){
+    public EstadoEnum getEstado() {
         return this.estado;
     }
 
-    public void setId(Integer id){
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public void setMonto(Double monto){
+    public void setMonto(BigDecimal monto) {
         this.monto = monto;
     }
 
-    public void setCuentaOrigen(Cuenta cuentaOrigen){
+    public void setCuentaOrigen(Cuenta cuentaOrigen) {
         this.cuentaOrigen = cuentaOrigen;
     }
 
-    public void setCuentaDestino(Cuenta cuentaDestino){
+    public void setCuentaDestino(Cuenta cuentaDestino) {
         this.cuentaDestino = cuentaDestino;
     }
 
-    public void setFechaRegistro(Date fechaRegistro){
+    public void setFechaRegistro(Date fechaRegistro) {
         this.fechaRegistro = fechaRegistro;
     }
 
-    public void setFechaProcesa(Date fechaProcesa){
+    public void setFechaProcesa(Date fechaProcesa) {
         this.fechaProcesa = fechaProcesa;
     }
 
-    public void setEstado(EstadoEnum estado){
+    public void setEstado(EstadoEnum estado) {
         this.estado = estado;
     }
 
     @Override
     public String toString() {
-        return "{" +
-                " monto ='" + getMonto() + "'" +
-                " cuentaOrigen ='" + getCuentaOrigen() + "'" +
-                " cuentaDestino ='" + getCuentaDestino() + "'" +
-                " fechaRegistro ='" + getFechaRegistro() + "'" +
-                " fechaProcesamiento ='" + getFechaProcesa() + 
-                "}";
+        return "{" + " monto ='" + getMonto() + "'" + " cuentaOrigen ='" + getCuentaOrigen() + "'" + " cuentaDestino ='"
+                + getCuentaDestino() + "'" + " fechaRegistro ='" + getFechaRegistro() + "'" + " fechaProcesamiento ='"
+                + getFechaProcesa() + "}";
     }
-    
+
 }
